Clean up WS initialization step 1
Moved the initialization code to its own file. State in the
WorkspaceServer class is vastly reduced.
