+++ src/us/kbase/workspace/WorkspaceServer.java
@@ -33,10 +33,7 @@ import static us.kbase.workspace.kbase.KBaseIdentifierFactory.processWorkspaceId
 import static us.kbase.workspace.kbase.KBasePermissions.translatePermission;
 
 import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
-import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -54,16 +51,9 @@ import ch.qos.logback.core.AppenderBase;
 
 //import org.apache.commons.lang3.builder.ToStringBuilder;
 
-import us.kbase.abstracthandle.AbstractHandleClient;
-import us.kbase.auth.AuthConfig;
 import us.kbase.auth.AuthException;
 import us.kbase.auth.AuthService;
-import us.kbase.auth.ConfigurableAuthService;
 import us.kbase.auth.RefreshingToken;
-import us.kbase.common.mongo.exceptions.InvalidHostException;
-import us.kbase.common.mongo.exceptions.MongoAuthException;
-import us.kbase.common.service.ServerException;
-import us.kbase.handlemngr.HandleMngrClient;
 import us.kbase.typedobj.core.AbsoluteTypeDefId;
 import us.kbase.typedobj.core.TempFilesManager;
 import us.kbase.typedobj.core.TypeDefId;
@@ -72,24 +62,21 @@ import us.kbase.typedobj.db.FuncDetailedInfo;
 import us.kbase.typedobj.db.ModuleDefId;
 import us.kbase.typedobj.db.TypeChange;
 import us.kbase.typedobj.db.TypeDetailedInfo;
-import us.kbase.typedobj.exceptions.TypeStorageException;
 import us.kbase.workspace.database.ByteArrayFileCacheManager.ByteArrayFileCache;
 import us.kbase.workspace.database.ResourceUsageConfigurationBuilder.ResourceUsageConfiguration;
 import us.kbase.workspace.database.ObjectChain;
-import us.kbase.workspace.database.ResourceUsageConfigurationBuilder;
 import us.kbase.workspace.database.SubObjectIdentifier;
 import us.kbase.workspace.database.Workspace;
-import us.kbase.workspace.database.WorkspaceDatabase;
 import us.kbase.workspace.database.ObjectIdentifier;
 import us.kbase.workspace.database.Permission;
 import us.kbase.workspace.database.WorkspaceIdentifier;
 import us.kbase.workspace.database.WorkspaceInformation;
 import us.kbase.workspace.database.WorkspaceObjectData;
 import us.kbase.workspace.database.WorkspaceUser;
-import us.kbase.workspace.database.exceptions.WorkspaceDBException;
-import us.kbase.workspace.database.mongo.MongoWorkspaceDB;
 import us.kbase.workspace.kbase.ArgUtils;
-import us.kbase.workspace.kbase.KBaseReferenceParser;
+import us.kbase.workspace.kbase.InitWorkspaceServer.InitReporter;
+import us.kbase.workspace.kbase.InitWorkspaceServer;
+import us.kbase.workspace.kbase.InitWorkspaceServer.WorkspaceInitResults;
 import us.kbase.workspace.kbase.WorkspaceAdministration;
 import us.kbase.workspace.kbase.WorkspaceServerMethods;
 //END_HEADER
@@ -127,51 +114,13 @@ public class WorkspaceServer extends JsonServerServlet {
     //TODO timestamps for startup script
     //TODO check shock version
     //TODO shock client should ignore extra fields
-    
-    //TODO make handle manager package?
 	
 	private static final String VER = "0.3.5";
 
-	//required deploy parameters:
-	private static final String HOST = "mongodb-host";
-	private static final String DB = "mongodb-database";
-	//startup workspace admin user
-	private static final String WSADMIN = "ws-admin";
-	//required backend param:
-	private static final String BACKEND_SECRET = "backend-secret"; 
-	//mongo db auth params:
-	private static final String USER = "mongodb-user";
-	private static final String PWD = "mongodb-pwd";
-	//mongo connection attempt limit
-	private static final String MONGO_RECONNECT = "mongodb-retry";
-	
-	//credentials to use for user queries
-	private static final String KBASE_ADMIN_USER = "kbase-admin-user";
-	private static final String KBASE_ADMIN_PWD = "kbase-admin-pwd";
-	
-	//handle service / manager info
-	private static final String IGNORE_HANDLE_SERVICE =
-			"ignore_handle_service";
-	private static final String HANDLE_SERVICE_URL = "handle-service-url";
-	private static final String HANDLE_MANAGER_URL = "handle-manager-url";
-	private static final String HANDLE_MANAGER_USER = "handle-manager-user";
-	private static final String HANDLE_MANAGER_PWD = "handle-manager-pwd";
-	private static final int TOKEN_REFRESH_INTERVAL = 24 * 60 * 60;
-	private static int maxUniqueIdCountPerCall = 100000;
-
-	//directory for temp files
-	private static final String TEMP_DIR = "temp-dir";
-	
-	
 	private static final long MAX_RPC_PACKAGE_SIZE = 1005000000;
 	private static final int MAX_RPC_PACKAGE_MEM_USE = 100000000;
 	
-	private final static int TOKEN_REFRESH_INTERVAL_SEC = 24 * 60 * 60;
-	
 	private static Map<String, String> wsConfig = null;
-	
-	private static int instanceCount = 0;
-	private static boolean wasTempFileCleaningDone = false;
 	
 	private final TempFilesManager tfm;
 	private final Workspace ws;
@@ -184,73 +133,9 @@ public class WorkspaceServer extends JsonServerServlet {
 	private ThreadLocal<Set<ByteArrayFileCache>> resourcesToDelete =
 			new ThreadLocal<Set<ByteArrayFileCache>>();
 	
-	private WorkspaceDatabase getDB(final String host, final String dbs,
-			final String secret, final String user, final String pwd,
-			final TempFilesManager tfm, final int mongoReconnectRetry) {
-		try {
-			if (user != null) {
-				return new MongoWorkspaceDB(host, dbs, secret, user, pwd, tfm,
-						mongoReconnectRetry);
-			} else {
-				return new MongoWorkspaceDB(host, dbs, secret, tfm,
-						mongoReconnectRetry);
-			}
-		} catch (UnknownHostException uhe) {
-			fail("Couldn't find mongo host " + host + ": " +
-					uhe.getLocalizedMessage());
-		} catch (IOException io) {
-			fail("Couldn't connect to mongo host " + host + ": " +
-					io.getLocalizedMessage());
-		} catch (MongoAuthException ae) {
-			fail("Not authorized: " + ae.getLocalizedMessage());
-		} catch (InvalidHostException ihe) {
-			fail(host + " is an invalid database host: "  +
-					ihe.getLocalizedMessage());
-		} catch (WorkspaceDBException uwde) {
-			fail("The workspace database is invalid: " +
-					uwde.getLocalizedMessage());
-		} catch (TypeStorageException tse) {
-			fail("There was a problem setting up the type storage system: " +
-					tse.getLocalizedMessage());
-		} catch (InterruptedException ie) {
-			fail("Connection to MongoDB was interrupted. This should never " +
-					"happen and indicates a programming problem. Error: " +
-					ie.getLocalizedMessage());
-		}
-		return null;
-	}
-	
-	private TempFilesManager initTempFilesManager() {
-		if (!wsConfig.containsKey(TEMP_DIR)) {
-			fail("Must provide param " + TEMP_DIR + " in config file");
-			return null;
-		}
-		try {
-			final TempFilesManager tfm = new TempFilesManager(
-					new File(wsConfig.get(TEMP_DIR)));
-			if (!wasTempFileCleaningDone) {
-				wasTempFileCleaningDone = true;
-				tfm.cleanup();
-			}
-			return tfm;
-		} catch (Exception e) {
-			fail(e.getLocalizedMessage());
-			return null;
-		}
-	}
-	
-	private void fail(final String error) {
-		logErr(error);
-		System.err.println(error);
-		startupFailed();
-	}
 	
 	public static void clearConfigForTests() {
 		wsConfig = null;
-	}
-	
-	public static void setMaximumUniqueIdCountForTests(final int count) {
-		maxUniqueIdCountPerCall = count;
 	}
 	
 	@Override
@@ -258,6 +143,10 @@ public class WorkspaceServer extends JsonServerServlet {
 		return ws.getTempFilesManager().generateTempFile("rpc", "json");
 	}
 	
+	public TempFilesManager getTempFilesManager() {
+		return tfm;
+	}
+
 	@Override
 	protected void onRpcMethodDone() {
 		if (resourcesToDelete.get() != null &&
@@ -268,10 +157,6 @@ public class WorkspaceServer extends JsonServerServlet {
 				} catch (Exception ignore) {}
 			resourcesToDelete.set(null);
 		}
-	}
-	
-	public TempFilesManager getTempFilesManager() {
-		return tfm;
 	}
 	
 	public ResourceUsageConfiguration getWorkspaceResourceUsageConfig() {
@@ -315,129 +200,26 @@ public class WorkspaceServer extends JsonServerServlet {
 		kbaseRootLogger.addAppender(kbaseAppender);
 	}
 	
-	private int getReconnectCount() {
-		final String rec = wsConfig.get(MONGO_RECONNECT);
-		Integer recint = null;
-		try {
-			recint = Integer.parseInt(rec); 
-		} catch (NumberFormatException nfe) {
-			//do nothing
-		}
-		if (recint == null) {
-			logInfo("Couldn't parse MongoDB reconnect value to an integer: " +
-					rec + ", using 0");
-			recint = 0;
-		} else if (recint < 0) {
-			logInfo("MongoDB reconnect value is < 0 (" + recint + "), using 0");
-			recint = 0;
-		} else {
-			logInfo("MongoDB reconnect value is " + recint);
-		}
-		return recint;
-	}
-	
+	private class WorkspaceInitReporter implements InitReporter {
 
-	private URL getHandleUrl(String configKey) {
-		final String urlStr = wsConfig.get(configKey);
-		if (urlStr == null || urlStr.isEmpty()) {
-			fail("Must provide param " + configKey + " in config file");
-			return null;
+		public boolean failed = false;
+		
+		@Override
+		public void reportInfo(final String info) {
+			logInfo(info);
+			System.out.println(info);
 		}
-		try {
-			return new URL(urlStr);
-		} catch (MalformedURLException e) {
-			fail("Invalid url for parameter " + configKey + ": " + urlStr);
-		}
-		return null;
-	}
-	
-	private RefreshingToken getHandleToken() {
-		final String user = wsConfig.get(HANDLE_MANAGER_USER);
-		final String pwd =  wsConfig.get(HANDLE_MANAGER_PWD);
-		//TODO make a method that checks for all required params and fails
-		if (user == null || user.isEmpty() || pwd == null || pwd.isEmpty()) {
-			fail("Must provide params " + HANDLE_MANAGER_USER + " and " +
-					HANDLE_MANAGER_PWD + " in config file");
-			return null;
-		}
-		try {
-			return AuthService.getRefreshingToken(
-					user, pwd, TOKEN_REFRESH_INTERVAL);
-		} catch (AuthException e) {
-			fail("Couldn't log in with handle manager credentials for user " +
-					user + ": " + e.getLocalizedMessage());
-		} catch (IOException e) {
-			fail("Couldn't contact the auth service to obtain a token for the handle manager: "
-					+ e.getLocalizedMessage());
-		}
-		return null;
-	}
-	
 
-	private boolean checkHandleServiceConnection(final URL handleServiceUrl) {
-		try {
-			final AbstractHandleClient cli = new AbstractHandleClient(
-					handleServiceUrl, handleMgrToken.getToken());
-			if (handleServiceUrl.getProtocol().equals("http")) {
-				System.out.println("Warning - the Handle Service url uses insecure http. https is recommended.");
-				logInfo("Warning - the Handle Service url uses insecure http. https is recommended.");
-				cli.setIsInsecureHttpConnectionAllowed(true);
-			}
-			cli.areReadable(new LinkedList<String>());
-		} catch (Exception e) {
-			if (!(e instanceof ServerException) ||
-					!e.getMessage().contains(
-							"can not execute select * from Handle")) {
-				fail("Could not establish a connection to the Handle Service at "
-						+ handleServiceUrl + ": " + e.getMessage());
-				return true;
-			}
+		@Override
+		public void reportFail(final String fail) {
+			logErr(fail);
+			System.err.println(fail);
+			startupFailed();
+			failed = true;
 		}
-		return false;
+		
 	}
 	
-	private boolean checkHandleManagerConnection() {
-		try {
-			final HandleMngrClient cli = new HandleMngrClient(
-					handleManagerUrl, handleMgrToken.getToken());
-			if (handleManagerUrl.getProtocol().equals("http")) {
-				System.out.println("Warning - the Handle Manager url uses insecure http. https is recommended.");
-				logInfo("Warning - the Handle Manager url uses insecure http. https is recommended.");
-				cli.setIsInsecureHttpConnectionAllowed(true);
-			}
-			cli.addReadAcl(Arrays.asList("FAKEHANDLE_-100"), "fakeuser");
-		} catch (Exception e) {
-			if (!(e instanceof ServerException) ||
-					!e.getMessage().contains(
-							"Unable to set acl(s) on handles FAKEHANDLE_-100")) {
-				fail("Could not establish a connection to the Handle Manager Service at "
-						+ handleManagerUrl + ": " + e.getMessage());
-				return true;
-			}
-		}
-		return false;
-	}
-	
-	private ConfigurableAuthService setUpAuthClient(
-			final String kbaseAdminUser,
-			final String kbaseAdminPwd) {
-		AuthConfig c = new AuthConfig();
-		ConfigurableAuthService auth;
-		try {
-			auth = new ConfigurableAuthService(c);
-			c.withRefreshingToken(auth.getRefreshingToken(
-					kbaseAdminUser, kbaseAdminPwd,
-					TOKEN_REFRESH_INTERVAL_SEC));
-			return auth;
-		} catch (AuthException e) {
-			fail("Couldn't log in the KBase administrative user " +
-					kbaseAdminUser + " : " + e.getLocalizedMessage());
-		} catch (IOException e) {
-			fail("Couldn't connect to authorization service at " +
-					c.getAuthServerURL() + " : " + e.getLocalizedMessage());
-		}
-		return null;
-	}
     //END_CLASS_HEADER
 
     public WorkspaceServer() throws Exception {
@@ -453,126 +235,26 @@ public class WorkspaceServer extends JsonServerServlet {
 			wsConfig = new HashMap<String, String>();
 			wsConfig.putAll(super.config);
 		}
-		tfm = initTempFilesManager();
-		boolean failed = tfm == null;
-		final String host = wsConfig.get(HOST);
-		if (host == null || host.isEmpty()) {
-			fail("Must provide param " + HOST + " in config file");
-			failed = true;
-		}
-		final String dbs = wsConfig.get(DB);
-		if (dbs == null || dbs.isEmpty()) {
-			fail("Must provide param " + DB + " in config file");
-			failed = true;
-		}
-		final String secret = wsConfig.get(BACKEND_SECRET);
-		if (secret == null || secret.isEmpty()) {
-			failed = true;
-			fail("Must provide param " + BACKEND_SECRET + " in config file");
-		}
-		final String user = wsConfig.get(USER);
-		final String pwd = wsConfig.get(PWD);
-		final boolean hasUser = user != null && !user.isEmpty();
-		final boolean hasPwd = pwd != null && !pwd.isEmpty();
 		
-		if (hasUser ^ hasPwd) {
-			fail(String.format("Must provide both %s and %s ",
-					USER, PWD) + "params in config file if authentication " + 
-					"is to be used");
-			failed = true;
-		}
-		final String ignoreHandle = wsConfig.get(IGNORE_HANDLE_SERVICE);
-		final boolean ignoreHandleService = 
-				ignoreHandle != null && !ignoreHandle.isEmpty();
-		URL handleServiceUrl = null;
-		if (ignoreHandleService) {
-			logInfo("Ignoring Handle Service config. Objects with handle IDs will fail typechecking.");
-			System.out.println("Ignoring Handle Service config. Objects with handle IDs will fail typechecking.");
-			handleManagerUrl = null;
-			handleMgrToken = null;
-		} else {
-			handleServiceUrl = getHandleUrl(HANDLE_SERVICE_URL);
-			failed = failed || handleServiceUrl == null;
-			handleManagerUrl = getHandleUrl(HANDLE_MANAGER_URL);
-			failed = failed || handleManagerUrl == null;
-			handleMgrToken = getHandleToken();
-			failed = failed || handleMgrToken == null;
-			if (!failed) {
-				failed = checkHandleServiceConnection(handleServiceUrl);
-			}
-			if (!failed) {
-				failed = checkHandleManagerConnection();
-			}
-		}
+		final WorkspaceInitReporter rep = new WorkspaceInitReporter();
 		
-		if (!wsConfig.containsKey(KBASE_ADMIN_USER)) {
-			fail("Must provide param " + KBASE_ADMIN_USER + " in config file");
-			failed = true;
-		}
-		final String adminUser = wsConfig.get(KBASE_ADMIN_USER);
-		if (!wsConfig.containsKey(KBASE_ADMIN_PWD)) {
-			fail("Must provide param " + KBASE_ADMIN_PWD + " in config file");
-			failed = true;
-		}
-		final String adminPwd = wsConfig.get(KBASE_ADMIN_PWD);
+		final WorkspaceInitResults res =
+				InitWorkspaceServer.initWorkspaceServer(wsConfig, rep);
 		
-		if (failed) {
-			fail("Server startup failed - all calls will error out.");
+		if (rep.failed) {
+			tfm = null;
 			ws = null;
 			wsmeth = null;
 			wsadmin = null;
+			handleManagerUrl = null;
+			handleMgrToken = null;
 		} else {
-			String params = "";
-			final List<String> paramSet = new LinkedList<String>(
-					Arrays.asList(HOST, DB, USER));
-			if (!ignoreHandleService) {
-				paramSet.addAll(Arrays.asList(HANDLE_SERVICE_URL,
-						HANDLE_MANAGER_URL, HANDLE_MANAGER_USER));
-			}
-			for (final String s: paramSet) {
-				if (wsConfig.containsKey(s)) {
-					params += s + "=" + wsConfig.get(s) + "\n";
-				}
-			}
-			params += BACKEND_SECRET + "=[redacted for your safety and comfort]\n";
-			if (pwd != null && !pwd.isEmpty()) {
-				params += PWD + "=[redacted for your safety and comfort]\n";
-			}
-			System.out.println("Starting server using connection parameters:\n"
-					+ params);
-			logInfo("Starting server using connection parameters:\n" + params);
-			System.out.println("Temporary file location: "
-					+ tfm.getTempDir());
-			logInfo("Temporary file location: " + tfm.getTempDir());
-			final int mongoConnectRetry = getReconnectCount();
-			final WorkspaceDatabase db = getDB(host, dbs, secret, user, pwd,
-					tfm, mongoConnectRetry);
-			if (db == null) {
-				fail("Server startup failed - all calls will error out.");
-				ws = null;
-				wsmeth = null;
-				wsadmin = null;
-			} else {
-				System.out.println(String.format("Initialized %s backend",
-						db.getBackendType()));
-				logInfo(String.format("Initialized %s backend",
-						db.getBackendType()));
-				ws = new Workspace(db,
-						new ResourceUsageConfigurationBuilder().build(),
-						new KBaseReferenceParser());
-				wsmeth = new WorkspaceServerMethods(ws, handleServiceUrl,
-						maxUniqueIdCountPerCall,
-						setUpAuthClient(adminUser, adminPwd));
-				wsadmin = new WorkspaceAdministration(ws, wsmeth,
-						wsConfig.get(WSADMIN));
-				final String mem = String.format(
-						"Started workspace server instance %s. Free mem: %s Total mem: %s, Max mem: %s",
-						++instanceCount, Runtime.getRuntime().freeMemory(),
-						Runtime.getRuntime().totalMemory(),
-						Runtime.getRuntime().maxMemory());
-				System.out.println(mem);
-				logInfo(mem);
-			}
+			tfm = res.getTempFilesManager();
+			ws = res.getWs();
+			wsmeth = res.getWsmeth();
+			wsadmin = res.getWsAdmin();
+			handleManagerUrl = res.getHandleManagerUrl();
+			handleMgrToken = res.getHandleMgrToken();
 			setRpcDiskCacheTempDir(tfm.getTempDir());
 		}
         //END_CONSTRUCTOR
